Actually updated for 1.7.9!

+++ src/main/java/com/onarandombox/multiverseinventories/api/share/Sharables.java
@@ -156,16 +156,16 @@ public final class Sharables implements Shares {
     /**
      * Sharing Health.
      */
-    public static final Sharable<Integer> HEALTH = new Sharable.Builder<Integer>("hit_points", Integer.class,
-            new SharableHandler<Integer>() {
+    public static final Sharable<Double> HEALTH = new Sharable.Builder<Double>("hit_points", Double.class,
+            new SharableHandler<Double>() {
                 @Override
                 public void updateProfile(PlayerProfile profile, Player player) {
-                    profile.set(HEALTH, player.getHealth());
+                    profile.set(HEALTH, (double) player.getHealth());
                 }
 
                 @Override
                 public boolean updatePlayer(Player player, PlayerProfile profile) {
-                    Integer value = profile.get(HEALTH);
+                    Double value = profile.get(HEALTH);
                     if (value == null) {
                         player.setHealth(PlayerStats.HEALTH);
                         return false;
